+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 * Only direct unconstrained dependencies decrease the health score.
 
-* Remove some of the `pubspec.lock` validation.
+* Removed superfluous `pubspec.lock` validation.
+
+* Recommend descriptions between 60 and 180 characters.
 
 ## 0.8.2