+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Only direct unconstrained dependencies decrease the health score.
 
+* Removed superfluous `pubspec.lock` validation.
+
 * Recommend descriptions between 60 and 180 characters.
 
 ## 0.8.2